Container_B/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
pika==1.3.2
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
alembic==1.12.1 